Created new wishing well scene file with recent updates
Copied the whitebox scene and added an instance of the wishing well level prefab to it to circumvent all errors  related to event listeners, etc. Also made the dash values harder to hit, re-fixed sideways checkpoint text, and changed the collectables number stated in the tutorial text in this new wishing well level.